Fix printf-format errors

+++ src/client/cl_cgame.cpp
@@ -1319,7 +1319,7 @@ Ghoul2 Insert End
 		}
 	}
 
-	Com_Error( ERR_DROP, "Bad cgame system trap: %i", args[0] );
+	Com_Error( ERR_DROP, "Bad cgame system trap: %lli", (long long int)args[0] );
 	return 0;
 }
