+++ src/client/cl_ui.cpp
@@ -1175,7 +1175,7 @@ Ghoul2 Insert End
 		}
 	}
 
-	Com_Error( ERR_DROP, "Bad UI system trap: %i", args[0] );
+	Com_Error( ERR_DROP, "Bad UI system trap: %lli", (long long int)args[0] );
 	return 0;
 }
