+++ src/server/sv_game.cpp
@@ -1105,7 +1105,7 @@ intptr_t SV_GameSystemCalls( intptr_t *args ) {
 		}
 	}
 
-	Com_Error( ERR_DROP, "Bad game system trap: %i", args[0] );
+	Com_Error( ERR_DROP, "Bad game system trap: %lli", (long long int)args[0] );
 	return -1;
 }
